permission export import

+++ app/Livewire/User/PermissionIndex.php
@@ -2,15 +2,23 @@
 
 namespace App\Livewire\User;
 
+use App\Exports\PermissionExport;
+use App\Imports\PermissionImport;
 use Livewire\Component;
 use Illuminate\Support\Str;
+use Livewire\WithFileUploads;
+use Maatwebsite\Excel\Facades\Excel;
+use RealRashid\SweetAlert\Facades\Alert;
 use Spatie\Permission\Models\Permission;
 
 class PermissionIndex extends Component
 {
+    use WithFileUploads;
     public $id, $name, $permissions;
     public $form = false;
     public $search = '';
+    public $formImport = 0;
+    public $fileImport;
     public function store()
     {
         $this->validate([
@@ -26,8 +34,12 @@ class PermissionIndex extends Component
     public function destroy($id)
     {
         $permission = Permission::find($id);
-        $permission->delete();
-        flash('Permission ' . $permission->name . ' deleted successfully.', 'success');
+        if ($permission->roles->count()) {
+            flash('Permission tidak bisa dihapus karena sedang dipakai', 'danger');
+        } else {
+            $permission->delete();
+            flash('Permission ' . $permission->name . ' deleted successfully.', 'success');
+        }
     }
     public function edit($id)
     {
@@ -52,14 +64,37 @@ class PermissionIndex extends Component
     {
         return view('components.placeholder.placeholder-text');
     }
-    public function mount()
+    public function export()
     {
-
+        try {
+            $time = now()->format('Y-m-d');
+            flash('Export successfully', 'success');
+            return Excel::download(new PermissionExport, 'permission_backup_' . $time . '.xlsx');
+        } catch (\Throwable $th) {
+            flash('Mohon maaf ' . $th->getMessage(), 'danger');
+        }
+    }
+    public function openFormImport()
+    {
+        $this->formImport = $this->formImport ?  0 : 1;
+    }
+    public function import()
+    {
+        try {
+            $this->validate([
+                'fileImport' => 'required|mimes:xlsx'
+            ]);
+            Excel::import(new PermissionImport, $this->fileImport->getRealPath());
+            Alert::success('Success', 'User imported successfully');
+            return redirect()->route('role-permission');
+        } catch (\Throwable $th) {
+            flash('Mohon maaf ' . $th->getMessage(), 'danger');
+        }
     }
     public function render()
     {
         $search = '%' . $this->search . '%';
-        $this->permissions = Permission::orderBy('name', 'asc')
+        $this->permissions = Permission::with(['roles'])->orderBy('name', 'asc')
             ->where('name', 'like', $search)
             ->get();
         return view('livewire.user.permission-index');
